@@ -19,15 +19,17 @@
   },
   "devDependencies": {
     "@types/jquery": "^3.3.31",
+    "node-polyfill-webpack-plugin": "^1.1.4",
     "source-map-loader": "^0.2.4",
     "ts-loader": "^6.2.1",
     "typescript": "^3.7.3",
-    "webpack": "^5.52.0",
-    "webpack-cli": "^4.8.0"
+    "webpack": "^5.59.1",
+    "webpack-cli": "^4.9.1"
   },
   "dependencies": {
     "aws-crt": "file:../../../",
     "aws-sdk": "^2.585.0",
-    "jquery": "^3.5.0"
+    "jquery": "^3.5.0",
+    "util": "^0.12.4"
   }
 }
