@@ -1,6 +1,12 @@
-/**
+/*
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0.
+ */
+
+/**
+ *
+ * @packageDocumentation
+ * @module crypto
  */
 
 import * as Crypto from "crypto-js";
@@ -10,24 +16,25 @@ import { TextEncoder } from "./polyfills";
 /**
  * Object that allows for continuous MD5 hashing of data.
  *
- * @module aws-crt
  * @category Crypto
  */
 export class Md5Hash {
     private hash?: Crypto.WordArray;
 
     /**
-     * Digests additional data
-     * @param data Additional data to digest
+     * Hashes additional data
+     * @param data Additional data to hash
      */
     update(data: Hashable) {
         this.hash = Crypto.MD5(data.toString(), this.hash ? this.hash.toString() : undefined);
     }
 
     /**
-     * Completes the hash computation and returns the final digest.
+     * Completes the hash computation and returns the final hash digest.
      *
      * @param truncate_to The maximum number of bytes to receive. Leave as undefined or 0 to receive the entire digest.
+     *
+     * @returns the final hash digest
      */
     finalize(truncate_to?: number): DataView {
         const digest = this.hash ? this.hash.toString() : '';
@@ -46,7 +53,8 @@ export class Md5Hash {
  * @param data The data to hash
  * @param truncate_to The maximum number of bytes to receive. Leave as undefined or 0 to receive the entire digest.
  *
- * @module aws-crt
+ * @returns the data's hash digest
+ *
  * @category Crypto
  */
 export function hash_md5(data: Hashable, truncate_to?: number): DataView {
@@ -58,25 +66,26 @@ export function hash_md5(data: Hashable, truncate_to?: number): DataView {
 /**
  * Object that allows for continuous SHA256 hashing of data.
  *
- * @module aws-crt
  * @category Crypto
  */
 export class Sha256Hash {
     private hash?: Crypto.WordArray;
 
     /**
-     * Digests additional data
-     * @param data Additional data to digest
+     * Hashes additional data
+     * @param data Additional data to hash
      */
     update(data: Hashable) {
         this.hash = Crypto.SHA256(data.toString(), this.hash ? this.hash.toString() : undefined);
     }
 
     /**
-    * Completes the hash computation and returns the final digest.
-    *
-    * @param truncate_to The maximum number of bytes to receive. Leave as undefined or 0 to receive the entire digest.
-    */
+     * Completes the hash computation and returns the final hash digest.
+     *
+     * @param truncate_to The maximum number of bytes to receive. Leave as undefined or 0 to receive the entire digest.
+     *
+     * @returns the final hash digest
+     */
     finalize(truncate_to?: number): DataView {
         const digest = this.hash ? this.hash.toString() : '';
         const truncated = digest.substring(0, truncate_to ? truncate_to : digest.length);
@@ -93,7 +102,8 @@ export class Sha256Hash {
  * @param data The data to hash
  * @param truncate_to The maximum number of bytes to receive. Leave as undefined or 0 to receive the entire digest.
  *
- * @module aws-crt
+ * @returns the data's hash digest
+ *
  * @category Crypto
  */
 export function hash_sha256(data: Hashable, truncate_to?: number): DataView {
@@ -107,25 +117,26 @@ export function hash_sha256(data: Hashable, truncate_to?: number): DataView {
 /**
  * Object that allows for continuous SHA1 hashing of data.
  *
- * @module aws-crt
  * @category Crypto
  */
  export class Sha1Hash {
     private hash?: Crypto.WordArray;
 
     /**
-     * Digests additional data
-     * @param data Additional data to digest
+     * Hashes additional data
+     * @param data Additional data to hash
      */
     update(data: Hashable) {
         this.hash = Crypto.SHA1(data.toString(), this.hash ? this.hash.toString() : undefined);
     }
 
     /**
-    * Completes the hash computation and returns the final digest.
-    *
-    * @param truncate_to The maximum number of bytes to receive. Leave as undefined or 0 to receive the entire digest.
-    */
+     * Completes the hash computation and returns the final hash digest.
+     *
+     * @param truncate_to The maximum number of bytes to receive. Leave as undefined or 0 to receive the entire digest.
+     *
+     * @returns the final hash digest
+     */
     finalize(truncate_to?: number): DataView {
         const digest = this.hash ? this.hash.toString() : '';
         const truncated = digest.substring(0, truncate_to ? truncate_to : digest.length);
@@ -142,7 +153,8 @@ export function hash_sha256(data: Hashable, truncate_to?: number): DataView {
  * @param data The data to hash
  * @param truncate_to The maximum number of bytes to receive. Leave as undefined or 0 to receive the entire digest.
  *
- * @module aws-crt
+ * @returns the data's hash digest
+ *
  * @category Crypto
  */
 export function hash_sha1(data: Hashable, truncate_to?: number): DataView {
@@ -156,30 +168,35 @@ export function hash_sha1(data: Hashable, truncate_to?: number): DataView {
 /**
  * Object that allows for continuous hashing of data with an hmac secret.
  *
- * @module aws-crt
  * @category Crypto
  */
 export class Sha256Hmac {
     private hmac: any;
 
+    /**
+     * Constructor for the Sha256Hmac class type
+     * @param secret secret key to seed the hmac process with
+     */
     constructor(secret: Hashable) {
         // @ts-ignore types file doesn't have this signature of create()
         this.hmac = Crypto.algo.HMAC.create(Crypto.algo.SHA256, secret);
     }
 
     /**
-     * Digests additional data
-     * @param data Additional data to digest
+     * Hashes additional data
+     * @param data Additional data to hash
      */
     update(data: Hashable) {
         this.hmac.update(data.toString());
     }
 
     /**
-    * Completes the hash computation and returns the final digest.
-    *
-    * @param truncate_to The maximum number of bytes to receive. Leave as undefined or 0 to receive the entire digest.
-    */
+     * Completes the hash computation and returns the final hmac digest.
+     *
+     * @param truncate_to The maximum number of bytes to receive. Leave as undefined or 0 to receive the entire digest.
+     *
+     * @returns the final hmac digest
+     */
     finalize(truncate_to?: number): DataView {
         const digest = this.hmac.finalize();
         const truncated = digest.toString().substring(0, truncate_to ? truncate_to : digest.length);
@@ -197,7 +214,8 @@ export class Sha256Hmac {
  * @param data The data to hash
  * @param truncate_to The maximum number of bytes to receive. Leave as undefined or 0 to receive the entire digest.
  *
- * @module aws-crt
+ * @returns the data's hmac digest
+ *
  * @category Crypto
  */
 export function hmac_sha256(secret: Hashable, data: Hashable, truncate_to?: number): DataView {
