@@ -1,27 +1,49 @@
-/**
+/*
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0.
  */
 
+/**
+ * @packageDocumentation
+ * @module mqtt
+ */
+
 import * as mqtt from "mqtt";
 import * as WebsocketUtils from "./ws";
+import * as auth from "./auth";
 import { Trie, TrieOp, Node as TrieNode } from "./trie";
 
 import { BufferedEventEmitter } from "../common/event";
 import { CrtError } from "../browser";
 import { ClientBootstrap, SocketOptions } from "./io";
-import { QoS, Payload, MqttRequest, MqttSubscribeRequest, MqttWill, OnMessageCallback } from "../common/mqtt";
+import {
+    QoS,
+    Payload,
+    MqttRequest,
+    MqttSubscribeRequest,
+    MqttWill,
+    OnMessageCallback,
+    MqttConnectionConnected,
+    MqttConnectionDisconnected,
+    MqttConnectionError,
+    MqttConnectionInterrupted,
+    MqttConnectionResumed
+} from "../common/mqtt";
 export { QoS, Payload, MqttRequest, MqttSubscribeRequest, MqttWill } from "../common/mqtt";
 
-/** @category MQTT */
+/**
+ * @category MQTT
+ */
 export type WebsocketOptions = WebsocketUtils.WebsocketOptions;
-/** @category MQTT */
-export type AWSCredentials = WebsocketUtils.AWSCredentials;
+
+/**
+ * @category MQTT
+ */
+export type AWSCredentials = auth.AWSCredentials;
 
 /**
  * Configuration options for an MQTT connection
  *
- * @module aws-crt
  * @category MQTT
  */
 export interface MqttConnectionConfig {
@@ -30,12 +52,16 @@ export interface MqttConnectionConfig {
     * If an ID is already in use, the other client will be disconnected.
     */
     client_id: string;
+
     /** Server name to connect to */
     host_name: string;
+
     /** Server port to connect to */
     port: number;
+
     /** Socket options, ignored in browser */
     socket_options: SocketOptions;
+
     /**
      * Whether or not to start a clean session with each reconnect.
      * If True, the server will forget all subscriptions with each reconnect.
@@ -47,13 +73,15 @@ export interface MqttConnectionConfig {
      * and sends mesages (with QoS1 or higher) that were published while the client was offline.
      */
     clean_session?: boolean;
+
     /**
      * The keep alive value, in seconds, to send in CONNECT packet.
      * A PING will automatically be sent at this interval.
      * The server will assume the connection is lost if no PING is received after 1.5X this value.
-     * This duration must be longer than {@link timeout}.
+     * This duration must be longer than {@link ping_timeout}.
      */
     keep_alive?: number;
+
     /**
      * Milliseconds to wait for ping response before client assumes
      * the connection is invalid and attempts to reconnect.
@@ -63,6 +91,7 @@ export interface MqttConnectionConfig {
      * but keep-alive options may not work the same way on every platform and OS version.
      */
     ping_timeout?: number;
+
     /**
      * Milliseconds to wait for the response to the operation requires response by protocol.
      * Set to zero to disable timeout. Otherwise, the operation will fail if no response is
@@ -70,17 +99,22 @@ export interface MqttConnectionConfig {
      * It applied to PUBLISH (QoS>0) and UNSUBSCRIBE now.
      */
     protocol_operation_timeout?: number;
+
     /**
      * Will to send with CONNECT packet. The will is
      * published by the server when its connection to the client is unexpectedly lost.
      */
     will?: MqttWill;
+
     /** Username to connect with */
     username?: string;
+
     /** Password to connect with */
     password?: string;
+
     /** Options for the underlying websocket connection */
     websocket?: WebsocketOptions;
+
     /** AWS credentials, which will be used to sign the websocket request */
     credentials?: AWSCredentials;
 }
@@ -88,13 +122,13 @@ export interface MqttConnectionConfig {
 /**
  * MQTT client
  *
- * @module aws-crt
  * @category MQTT
  */
 export class MqttClient {
     constructor(bootstrap?: ClientBootstrap) {
 
     }
+
     /**
      * Creates a new {@link MqttClientConnection}
      * @param config Configuration for the connection
@@ -176,7 +210,6 @@ function normalize_payload(payload: Payload): Buffer | string {
 /**
  * MQTT client connection
  *
- * @module aws-crt
  * @category MQTT
  */
 export class MqttClientConnection extends BufferedEventEmitter {
@@ -228,35 +261,68 @@ export class MqttClientConnection extends BufferedEventEmitter {
         this.connection.on('end', this.on_disconnected);
     }
 
-    /** Emitted when the connection is ready and is about to start sending response data */
-    on(event: 'connect', listener: (session_present: boolean) => void): this;
-
-    /** Emitted when connection has closed sucessfully. */
-    on(event: 'disconnect', listener: () => void): this;
+    /**
+     * Emitted when the connection successfully establishes itself for the first time
+     *
+     * @param event the type of event (connect)
+     * @param listener the event listener to use
+     *
+     * @event
+     */
+    on(event: 'connect', listener: MqttConnectionConnected): this;
 
     /**
-     * Emitted when an error occurs
-     * @param error - A CrtError containing the error that occurred
+     * Emitted when connection has disconnected sucessfully.
+     *
+     * @param event the type of event (disconnect)
+     * @param listener the event listener to use
+     *
+     * @event
      */
-    on(event: 'error', listener: (error: CrtError) => void): this;
+    on(event: 'disconnect', listener: MqttConnectionDisconnected): this;
+
+    /**
+     * Emitted when an error occurs.  The error will contain the error
+     * code and message.
+     *
+     * @param event the type of event (error)
+     * @param listener the event listener to use
+     *
+     * @event
+     */
+    on(event: 'error', listener: MqttConnectionError): this;
 
     /**
      * Emitted when the connection is dropped unexpectedly. The error will contain the error
-     * code and message.
+     * code and message.  The underlying mqtt implementation will attempt to reconnect.
+     *
+     * @param event the type of event (interrupt)
+     * @param listener the event listener to use
+     *
+     * @event
      */
-    on(event: 'interrupt', listener: (error: CrtError) => void): this;
+    on(event: 'interrupt', listener: MqttConnectionInterrupted): this;
 
     /**
-     * Emitted when the connection reconnects. Only triggers on connections after the initial one.
+     * Emitted when the connection reconnects (after an interrupt). Only triggers on connections after the initial one.
+     *
+     * @param event the type of event (resume)
+     * @param listener the event listener to use
+     *
+     * @event
      */
-    on(event: 'resume', listener: (return_code: number, session_present: boolean) => void): this;
+    on(event: 'resume', listener: MqttConnectionResumed): this;
 
     /**
      * Emitted when any MQTT publish message arrives.
+     *
+     * @param event the type of event (message)
+     * @param listener the event listener to use
+     *
+     * @event
      */
     on(event: 'message', listener: OnMessageCallback): this;
 
-    /** @internal */
     on(event: string | symbol, listener: (...args: any[]) => void): this {
         return super.on(event, listener);
     }
