@@ -1,13 +1,22 @@
-/**
+/*
+ *
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0.
  */
 
 /**
+ *
+ * A module containing support for creating http connections and making requests on them.
+ *
+ * @packageDocumentation
+ * @module http
+ * @preferred
+ */
+
+/**
  * HTTP protocol version
  *
- * @module aws-crt
-* @category HTTP
+ * @category HTTP
  */
 export enum HttpVersion {
     Unknown = 0,
@@ -22,8 +31,7 @@ export enum HttpVersion {
 /**
  * Headers are exposed as 2 element arrays: [name, value]
  *
- * @module aws-crt
-* @category HTTP
+ * @category HTTP
  */
 export type HttpHeader = [string, string];
 
@@ -91,21 +99,36 @@ export interface HttpHeaders {
 /**
  * Proxy authentication types
  *
- * @module aws-crt
  * @category HTTP
  */
 export enum HttpProxyAuthenticationType {
+
+    /**
+     * No to-proxy authentication logic
+     */
     None = 0,
+
+    /**
+     * Use basic authentication (user/pass).  Supply these values in {@link HttpProxyOptions}
+     */
     Basic = 1,
 };
 
 /**
  * Options used when connecting to an HTTP endpoint via a proxy
  *
- * @module aws-crt
  * @category HTTP
  */
-export class HttpProxyOptions {
+export class CommonHttpProxyOptions {
+
+    /**
+     *
+     * @param host_name endpoint of the proxy to use
+     * @param port port of proxy to use
+     * @param auth_method type of authentication to use with the proxy
+     * @param auth_username (basic authentication only) proxy username
+     * @param auth_password (basic authentication only) password associated with the username
+     */
     constructor(
         public host_name: string,
         public port: number,
@@ -115,3 +138,60 @@ export class HttpProxyOptions {
     ) {
     }
 }
+
+/**
+ * Listener signature for event emitted from an {@link HttpClientConnection} when the connection reaches the
+ * connected state
+ *
+ * @asMemberOf HttpClientConnection
+ * @category HTTP
+ */
+export type HttpClientConnectionConnected = () => void;
+
+/**
+ * Listener signature for event emitted from an {@link HttpClientConnection} when an error occurs
+ *
+ * @param error - A CrtError containing the error that occurred
+ *
+ * @asMemberOf HttpClientConnection
+ * @category HTTP
+ */
+export type HttpClientConnectionError = (error: Error) => void;
+
+/**
+ * Listener signature for event emitted from an {@link HttpClientConnection} when the connection has been closed
+ *
+ * @asMemberOf HttpClientConnection
+ * @category HTTP
+ */
+export type HttpClientConnectionClosed = () => void;
+
+/**
+ * Listener signature for event emitted from an {@link HttpClientStream} when http response data is available
+ *
+ * @param body_data - The chunk of body data
+ *
+ * @asMemberOf HttpClientStream
+ * @category HTTP
+ */
+export type HttpStreamData = (body_data: ArrayBuffer) => void;
+
+
+/**
+ * Listener signature for event emitted from an {@link HttpClientStream} when an http stream error occurs
+ *
+ * @param error - A CrtError containing the error that occurred
+ *
+ * @asMemberOf HttpClientStream
+ * @category HTTP
+ */
+export type HttpStreamError = (error: Error) => void;
+
+/**
+ * Listener signature for event emitted from an {@link HttpClientStream} when the http stream has completed.
+ *
+ * @asMemberOf HttpClientStream
+ * @category HTTP
+ */
+export type HttpStreamComplete = () => void;
+
