@@ -1,7 +1,16 @@
-/**
+/*
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0.
  */
+
+/**
+ * Module for AWS IoT configuration and connection establishment
+ *
+ * @packageDocumentation
+ * @module aws-iot
+ * @preferred
+ */
+
 import { MqttConnectionConfig, MqttWill } from "./mqtt";
 import * as io from "./io";
 import * as platform from '../common/platform';
@@ -15,14 +24,37 @@ import {
     AwsSigningConfig
 } from "./auth";
 
-/** @category IoT */
+/**
+ * Websocket-specific mqtt connection configuration options
+ *
+ * @category IoT
+ */
 export interface WebsocketConfig {
+
+    /** Sources the AWS Credentials used to sign the websocket connection handshake */
     credentials_provider: AwsCredentialsProvider;
+
+    /**
+     * (Optional) factory function to create the configuration used to sign the websocket handshake.  Leave null
+     * to use the default settings.
+     */
     create_signing_config?: () => AwsSigningConfig;
 
+    /** (Optional) http proxy configuration */
     proxy_options?: HttpProxyOptions;
+
+    /** AWS region the websocket connection is being established in.  Must match the region embedded in the
+     * endpoint.
+     */
     region: string;
+
+    /**
+     * (Optional) override for the service name used in signing the websocket handshake.  Leave null to use the
+     * default (iotdevicegateway)
+     */
     service?: string;
+
+    /** (Optional)  TLS configuration to use when establishing the connection */
     tls_ctx_options?: io.TlsContextOptions;
 }
 
@@ -30,7 +62,6 @@ export interface WebsocketConfig {
  * Builder functions to create a {@link MqttConnectionConfig} which can then be used to create
  * a {@link MqttClientConnection}, configured for use with AWS IoT.
  *
- * @module aws-crt
  * @category IoT
  */
 export class AwsIotMqttConnectionConfigBuilder {
@@ -275,7 +306,7 @@ export class AwsIotMqttConnectionConfigBuilder {
     }
 
     /**
-     * Returns the configured MqttConnectionConfig.  On the first invocation of this function, the tls context is cached
+     * Returns the configured MqttConnectionConfig.  On the first invocation of this function, the TLS context is cached
      * and re-used on all subsequent calls to build().
      * @returns The configured MqttConnectionConfig
      */
@@ -285,8 +316,8 @@ export class AwsIotMqttConnectionConfigBuilder {
         }
 
         /*
-         * By caching and reusing the tls context we get an enormous memory savings on a per-connection basis.
-         * The tradeoff is that you can't modify tls options in between calls to build.
+         * By caching and reusing the TLS context we get an enormous memory savings on a per-connection basis.
+         * The tradeoff is that you can't modify TLS options in between calls to build.
          * Previously we were making a new one with every single connection which had a huge negative impact on large
          * scale tests.
          */
