@@ -1,10 +1,18 @@
-/**
+/*
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0.
  */
 
+// The only exported types are http-related.  If this changes, we'll need to rework the pseudo-module configuration
+// for documentation
+
+/**
+ * @packageDocumentation
+ * @module http
+ */
+
 import { InputStream } from "./io";
-import { AwsSigningAlgorithm, AwsSignatureType, AwsSignedBodyValue, AwsSignedBodyHeaderType } from "./auth";
+import { AwsSigningConfig } from "./auth";
 import { HttpHeader, HttpHeaders as CommonHttpHeaders } from "../common/http";
 import { OnMessageCallback, QoS } from "../common/mqtt";
 
@@ -14,7 +22,7 @@ import { OnMessageCallback, QoS } from "../common/mqtt";
  */
 type NativeHandle = any;
 
-/** @category System */
+/** @internal */
 type StringLike = string | ArrayBuffer | ArrayBufferView;
 
 /* common */
@@ -269,7 +277,7 @@ export function http_connection_manager_release(manager: NativeHandle, connectio
 /**
  * A collection of HTTP headers
  *
- * @module aws-crt
+ * @module http
  * @category HTTP
  */
 export class HttpHeaders implements CommonHttpHeaders {
@@ -341,6 +349,8 @@ export class HttpHeaders implements CommonHttpHeaders {
  * Definition for an outgoing HTTP request.
  *
  * The request may be transformed (ex: signing the request) before its data is eventually sent.
+ *
+ * @internal
  */
 export class HttpRequest {
     constructor(method: string, path: string, headers?: HttpHeaders, body?: InputStream);
@@ -361,69 +371,6 @@ export class AwsCredentialsProvider {
 
     static newDefault(bootstrap?: NativeHandle): AwsCredentialsProvider;
     static newStatic(access_key: StringLike, secret_key: StringLike, session_token?: StringLike): AwsCredentialsProvider;
-}
-
-/**
- * Configuration for use in AWS-related signing.
- * AwsSigningConfig is immutable.
- * It is good practice to use a new config for each signature, or the date might get too old.
- */
-export interface AwsSigningConfig {
-    /** Which signing process to invoke */
-    algorithm: AwsSigningAlgorithm;
-    /** What kind of signature to compute */
-    signature_type: AwsSignatureType;
-    /** Credentials provider to fetch signing credentials with */
-    provider: AwsCredentialsProvider;
-    /** The region to sign against */
-    region: string;
-    /** Name of service to sign a request for */
-    service?: string;
-    /**
-     * Date and time to use during the signing process. If not provided then
-     * the current time in UTC is used. Naive dates (lacking timezone info)
-     * are assumed to be in local time
-     */
-    date?: Date;
-    /**
-     * Headers to skip when signing.
-     *
-     * Skipping auth-required headers will result in an unusable signature.
-     * Headers injected by the signing process are not skippable.
-     * This function does not override the internal check function
-     * (x-amzn-trace-id, user-agent), but rather supplements it.
-     * In particular, a header will get signed if and only if it returns
-     * true to both the internal check (skips x-amzn-trace-id, user-agent)
-     * and is found in this list (if defined)
-     */
-    header_blacklist?: string[];
-    /**
-     * Set true to double-encode the resource path when constructing the
-     * canonical request. By default, all services except S3 use double encoding.
-     */
-    use_double_uri_encode?: boolean;
-    /**
-     * Whether the resource paths are normalized when building the canonical request.
-     */
-    should_normalize_uri_path?: boolean;
-    /**
-     * Should the session token be omitted from the signing process?  This should only be
-     * true when making a websocket handshake with IoT Core.
-     */
-    omit_session_token?: boolean;
-    /**
-     * Value to use as the canonical request's body value.
-     *
-     * Typically, this is the SHA-256 of the payload, written as lowercase hex.
-     * If this has been precalculated, it can be set here.
-     * Special values used by certain services can also be set (see {@link AwsSignedBodyValue}).
-     * If undefined (the default), the typical value will be calculated from the payload during signing.
-     */
-    signed_body_value?: string;
-    /** Controls what header, if any, should be added to the request, containing the body value */
-    signed_body_header?: AwsSignedBodyHeaderType;
-    /** Query param signing only: how long the pre-signed URL is valid for */
-    expiration_in_seconds?: number;
 }
 
 /** @internal */
