@@ -1,6 +1,11 @@
-/**
+/*
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0.
+ */
+
+/**
+ * @packageDocumentation
+ * @module crypto
  */
 
 import crt_native from './binding';
@@ -14,15 +19,15 @@ import { Hashable } from "../common/crypto";
  */
 abstract class Hash extends NativeResource {
     /**
-     * Digest additional data.
-     * @param data Additional data to digest
+     * Hash additional data.
+     * @param data Additional data to hash
      */
     update(data: Hashable) {
         crt_native.hash_update(this.native_handle(), data);
     }
 
     /**
-     * Completes the hash computation and returns the final digest.
+     * Completes the hash computation and returns the final hash digest.
      *
      * @param truncate_to The maximum number of bytes to receive. Leave as undefined or 0 to receive the entire digest.
      */
@@ -38,7 +43,6 @@ abstract class Hash extends NativeResource {
 /**
  * Object that allows for continuous MD5 hashing of data.
  *
- * @module aws-crt
  * @category Crypto
  */
 export class Md5Hash extends Hash {
@@ -54,7 +58,6 @@ export class Md5Hash extends Hash {
  * @param data The data to hash
  * @param truncate_to The maximum number of bytes to receive. Leave as undefined or 0 to receive the entire digest.
  *
- * @module aws-crt
  * @category Crypto
  */
 export function hash_md5(data: Hashable, truncate_to?: number): DataView {
@@ -64,7 +67,6 @@ export function hash_md5(data: Hashable, truncate_to?: number): DataView {
 /**
  * Object that allows for continuous SHA256 hashing of data.
  *
- * @module aws-crt
  * @category Crypto
  */
 export class Sha256Hash extends Hash {
@@ -81,7 +83,6 @@ export class Sha256Hash extends Hash {
  * @param data The data to hash
  * @param truncate_to The maximum number of bytes to receive. Leave as undefined or 0 to receive the entire digest.
  *
- * @module aws-crt
  * @category Crypto
  */
 export function hash_sha256(data: Hashable, truncate_to?: number): DataView {
@@ -91,7 +92,6 @@ export function hash_sha256(data: Hashable, truncate_to?: number): DataView {
 /**
  * Object that allows for continuous SHA1 hashing of data.
  *
- * @module aws-crt
  * @category Crypto
  */
  export class Sha1Hash extends Hash {
@@ -108,7 +108,6 @@ export function hash_sha256(data: Hashable, truncate_to?: number): DataView {
  * @param data The data to hash
  * @param truncate_to The maximum number of bytes to receive. Leave as undefined or 0 to receive the entire digest.
  *
- * @module aws-crt
  * @category Crypto
  */
 export function hash_sha1(data: Hashable, truncate_to?: number): DataView {
@@ -118,19 +117,20 @@ export function hash_sha1(data: Hashable, truncate_to?: number): DataView {
 /**
  * Object that allows for continuous hashing of data with an hmac secret.
  *
- * @module aws-crt
  * @category Crypto
  */
 abstract class Hmac extends NativeResource {
     /**
-     * Digest additional data.
+     * Hash additional data.
+     *
+     * @param data additional data to hash
      */
     update(data: Hashable) {
         crt_native.hmac_update(this.native_handle(), data);
     }
 
     /**
-     * Completes the hash computation and returns the final digest.
+     * Completes the hash computation and returns the final hmac digest.
      *
      * @param truncate_to The maximum number of bytes to receive. Leave as undefined or 0 to receive the entire digest.
      */
@@ -146,7 +146,6 @@ abstract class Hmac extends NativeResource {
 /**
  * Object that allows for continuous SHA256 HMAC hashing of data.
  *
- * @module aws-crt
  * @category Crypto
  */
 export class Sha256Hmac extends Hmac {
@@ -163,7 +162,6 @@ export class Sha256Hmac extends Hmac {
  * @param data The data to hash
  * @param truncate_to The maximum number of bytes to receive. Leave as undefined or 0 to receive the entire digest.
  *
- * @module aws-crt
  * @category Crypto
  */
 export function hmac_sha256(secret: Hashable, data: Hashable, truncate_to?: number): DataView {
